# specs/001-agentic-risk-manager/contracts/api-spec.yaml
openapi: 3.1.0
info:
  title: Crypto Risk Manager Core API
  version: 1.0.0
  description: High-performance quantitative risk engine, portfolio aggregation, and rebalance advisor API.
paths:
...
  /api/v1/risk/single:
    post:
      summary: Compute quantitative single-deal risk metrics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [asset_id, lookback_days]
              properties:
                asset_id: { type: string }
                lookback_days: { type: integer, default: 90 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_id: { type: string }
                  reference_index: { type: string, example: "BTC" }
                  beta: { type: number, description: "Beta against Bitcoin" }
                  delta: { type: number }
                  volatility_annualized: { type: number }
                  standard_deviation: { type: number }
                  sharpe_ratio: { type: number, description: "Sharpe ratio with Rf=0" }
                  treynor_ratio: { type: number, description: "Treynor ratio with Rf=0" }
                  math_engine_version: { type: string, example: "v1.1.0", description: "Deterministic math engine version per Constitution III" }
  /api/v1/risk/portfolio:
    post:
      summary: Compute aggregate Greeks, covariance matrix, and portfolio Sharpe & Treynor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet_address]
              properties:
                wallet_address: { type: string }
                lookback_days: { type: integer, default: 90 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  net_delta: { type: number }
                  net_gamma: { type: number }
                  net_vega: { type: number }
                  sharpe_ratio: { type: number }
                  treynor_ratio: { type: number }
                  assets: { type: array, items: { type: string } }
                  covariance_matrix:
                    type: array
                    items:
                      type: array
                      items: { type: number }
                  correlation_matrix:
                    type: array
                    items:
                      type: array
                      items: { type: number }
                  math_engine_version: { type: string, example: "v1.1.0", description: "Deterministic math engine version per Constitution III" }
  /api/v1/risk/report:
    post:
      summary: Generate comprehensive Risk Report with VaR, Expected Shortfall, and rebalance suggestions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet_address]
              properties:
                wallet_address: { type: string }
                confidence_levels: { type: array, items: { type: number }, default: [0.95, 0.99] }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id: { type: string, format: uuid }
                  timestamp: { type: string, format: date-time }
                  var_95_usd: { type: number }
                  var_99_usd: { type: number }
                  es_95_usd: { type: number }
                  es_99_usd: { type: number }
                  rebalancing_suggestions:
                    type: array
                    items:
                      $ref: '#/components/schemas/RebalanceAction'
                  math_engine_version: { type: string, example: "v1.1.0", description: "Deterministic math engine version per Constitution III" }
...
components:
  schemas:
...
    RebalanceAction:
      type: object
      properties:
        action_type: { type: string, enum: [trim, accumulate, hedge] }
        asset_id: { type: string }
        target_delta_usd: { type: number }
        rationale: { type: string }
        recommended_venue: { type: string, enum: [1inch, uniswap] }